backend/requirements.txt
Flask==3.0
Flask-SQLAlchemy==2.5.1
mysql-connector-python==8.0.23
python-dotenv==0.17.1
Werkzeug==2.0.2